+++ wireviz/__main__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 
 if __name__ == "__main__":
-    from wireviz.wireviz import main
-    main()
+    from .wireviz import main
+    main(prog_name="wireviz")