+++ changelog.md
@@ -14,6 +14,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2021-06-22
+
+### Added
+
+* Common library file now included in distribution
+* `-c`/`--common`/`--prepend-common-lib` option includes the common library.
+* `click` is now an installation dependency.
+
+### Changed
+
+* Implemented command line interface with `click` instead of `argparse`
+
+## 1.0.1 - 2021-06-21
+
+### Fixed
+
+* Entry point for console script corrected sire the wireviz executable worked
+
 ## 1.0.0 - 2021-06-21
 
 ### Added